+++ packages/app-page-builder/src/types.ts
@@ -391,8 +391,6 @@ export type PbEditorPageSettingsPlugin = Plugin & {
     description: string;
     /* Settings group icon */
     icon: ReactNode;
-    /* GraphQL query fields to include in the `settings` subselect */
-    fields: string;
     /* Render function that handles the specified `fields` */
     render: (params: {
         data: Record<string, any>;
@@ -401,6 +399,11 @@ export type PbEditorPageSettingsPlugin = Plugin & {
         Bind: BindComponent;
     }) => ReactNode;
 };
+
+export interface PbEditorPageQueryFieldsPlugin extends Plugin {
+    type: "pb-editor-page-query-fields";
+    fields: string;
+}
 
 export type PbIcon = {
     /**