+++ packages/app/src/graphql/index.ts
@@ -1,3 +1,0 @@
-export { default as createOmitTypenameLink } from "./createOmitTypenameLink";
-export { default as createSetContextLink } from "./createSetContextLink";
-export { default as createConsoleLink } from "./createConsoleLink";